html/dom/elements/global-attributes/the-translate-attribute-011.html

<!DOCTYPE html>
<html  lang="en" >
<head>
<meta charset="utf-8"/>
<title>translate-enabled state, yes inside no</title>
<link rel='author' title='Richard Ishida' href='mailto:ishida@w3.org'>
<link rel='help' href='http://www.w3.org/TR/html5/dom.html#the-translate-attribute'>
<script src="http://w3c-test.org/resources/testharness.js"></script>
<script src="http://w3c-test.org/resources/testharnessreport.js"></script>
<link rel="stylesheet" href="http://w3c-test.org/resource/testharness.css">
<meta name="assert" content="If the translate attribute is set to &quot;yes&quot; on an element inside an element with the translate attribute set to &quot;no&quot;, javascript will detect the translation mode of text in the inner element as translate-enabled.">
<style type='text/css'>
/* the CSS below is not part of the test */
		.test { border: 0; }
		.test div { color:white; padding: 5px; background-color: red; height: 90px; width: 90px; }

		</style>
</head>
<body>
<p class="instructions">Test passes if you see a green square.</p>


<div class="test"><div id="box" translate="no">Fail <span id="spantest" translate="yes">&nbsp;</span></div></div>


<script>pass = false; if (document.getElementById('spantest').translate == true) {  pass=true; }
if (pass) {  box.style.backgroundColor = '#00FF00';  box.style.color = 'white'; box.innerHTML = 'Pass'; box.style.height = '100px'; box.style.width = '100px'; }
test(function () {
    assert_true(pass);
	}, '');
</script><div id='log'></div></body>
</html>
